# docs/swagger.yaml
definitions:
...
  main.releaseOptions:
    properties:
      atomic:
        type: boolean
      cleanup_on_fail:
        type: boolean
      create_namespace:
        description: only install
        type: boolean
      dependency_update:
        type: boolean
      description:
        type: string
      devel:
        type: boolean
      disable_hooks:
        type: boolean
      dry_run:
        description: common
        type: boolean
      force:
        description: only upgrade
        type: boolean
      install:
        type: boolean
      recreate:
        type: boolean
      set:
        items:
          type: string
        type: array
      set_string:
        items:
          type: string
        type: array
      skip_crds:
        type: boolean
      sub_notes:
        type: boolean
      timeout:
        type: string
      values:
        type: string
      wait:
        type: boolean
    type: object
...
  main.respBody:
    properties:
      code:
        description: 0 or 1, 0 is ok, 1 is error
        type: integer
      data:
        type: object
      error:
        type: string
    type: object
info:
  contact:
    email: mika055@163.com
    name: polya
  description: This is a api proxy of helm.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: Helm API Proxy
  version: 0.0.1
paths:
...
  /namespaces/{namespace}/releases:
    get:
      description: 根据命名空间获取release信息列表(helm list)
      parameters:
      - description: release所在k8s的命名空间
        in: path
        name: namespace
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 获取helm的release列表
      tags:
      - Release
  /namespaces/{namespace}/releases/{release}:
    delete:
      description: 卸载chart的实例(helm uninstall)
      parameters:
      - description: release所在k8s的命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: release名称
        in: path
        name: release
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 卸载release
      tags:
      - Release
    get:
      description: 根据名称获取release信息(helm get)
      parameters:
      - description: release所在k8s的命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: release名称
        in: path
        name: release
        required: true
        type: string
      - description: Enums(all, hooks, manifest, notes, values)
        in: query
        name: info
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 获取release详细信息
      tags:
      - Release
    post:
      description: 安装chart的实例(helm install)
      parameters:
      - description: release所在k8s的命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: release名称
        in: path
        name: release
        required: true
        type: string
      - description: chart名称
        in: query
        name: chart
        required: true
        type: string
      - description: 安装可选项
        in: body
        name: options
        required: true
        schema:
          $ref: '#/definitions/main.releaseOptions'
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 安装release
      tags:
      - Release
    put:
      description: 升级release(helm upgrade)
      parameters:
      - description: release所在k8s的命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: release名称
        in: path
        name: release
        required: true
        type: string
      - description: chart名称
        in: query
        name: chart
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: release升级
      tags:
      - Release
  /namespaces/{namespace}/releases/{release}/histories:
    get:
      description: 获取release历史记录(helm release history)
      parameters:
      - description: release所在k8s的命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: release名称
        in: path
        name: release
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 查看release历史记录
      tags:
      - Release
  /namespaces/{namespace}/releases/{release}/status:
    get:
      description: 获取release状态信息(helm status)
      parameters:
      - description: release所在k8s的命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: release名称
        in: path
        name: release
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: 查看release状态
      tags:
      - Release
  /namespaces/{namespace}/releases/{release}/versions/{reversion}:
    put:
      description: 回滚release到之前版本(helm rollback)
      parameters:
      - description: release所在k8s的命名空间
        in: path
        name: namespace
        required: true
        type: string
      - description: release名称
        in: path
        name: release
        required: true
        type: string
      - description: chart版本号
        in: path
        name: versions
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/main.respBody'
      summary: release回滚
      tags:
      - Release
...
swagger: "2.0"